src/test/kotlin/advent/Day05Test.kt
package advent


import utils.getFileAsText
import kotlin.test.Test
import kotlin.test.assertEquals

class Day05Test {
    private val example = "data/05_ex.txt"
    private val input = "data/05_in.txt"

    private val service = Day05()

    @Test
    fun example1() {
        val puzzleContent = getFileAsText(example)
        val response = service.execute01(puzzleContent)
        assertEquals(35L, response)
    }

    @Test
    fun exercise1() {
        val puzzleContent = getFileAsText(input)
        val response = service.execute01(puzzleContent)
        println(response)
        assertEquals(84470622, response)
    }

    @Test
    fun example2() {
        val puzzleContent = getFileAsText(example)
        val response = service.execute02(puzzleContent)
        assertEquals(46, response)
    }

    @Test
    fun exercise2() {
        val puzzleContent = getFileAsText(input)
        val response = service.execute02(puzzleContent)
        println(response)
        assertEquals(26714516, response)
    }

}